+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 In this document, most significant changes are noted.
 ## [0.0.1 b1] - 2022.11.24
 ### Added
-* Logging: Log type 'alert'
 * ModbusTCP functionalities
+* fldsvsmanager as a manager for scheduled services
+* Logging: Log type 'alert'
 ### Fixed
 * MariaDB connector error to log
 ## [0.0.1] - 2022.11.21